+++ internal/api/auth/validators_test.go
@@ -431,17 +431,17 @@ func TestValidators(t *testing.T) {
 			runTests(tests, db)
 		})
 
-		Convey("When testing ValidateNodeQueueAccess", func() {
+		Convey("When testing ValidateDeviceQueueAccess", func() {
 			tests := []validatorTest{
 				{
 					Name:       "global admin users can read, list, update and delete",
-					Validators: []ValidatorFunc{ValidateNodeQueueAccess(devices[0].DevEUI, Create), ValidateNodeQueueAccess(devices[0].DevEUI, Read), ValidateNodeQueueAccess(devices[0].DevEUI, List), ValidateNodeQueueAccess(devices[0].DevEUI, Update), ValidateNodeQueueAccess(devices[0].DevEUI, Delete)},
+					Validators: []ValidatorFunc{ValidateDeviceQueueAccess(devices[0].DevEUI, Create), ValidateDeviceQueueAccess(devices[0].DevEUI, Read), ValidateDeviceQueueAccess(devices[0].DevEUI, List), ValidateDeviceQueueAccess(devices[0].DevEUI, Update), ValidateDeviceQueueAccess(devices[0].DevEUI, Delete)},
 					Claims:     Claims{Username: "user1"},
 					ExpectedOK: true,
 				},
 				{
 					Name:       "other users can not read, list, update and delete",
-					Validators: []ValidatorFunc{ValidateNodeQueueAccess(devices[0].DevEUI, Create), ValidateNodeQueueAccess(devices[0].DevEUI, Read), ValidateNodeQueueAccess(devices[0].DevEUI, List), ValidateNodeQueueAccess(devices[0].DevEUI, Update), ValidateNodeQueueAccess(devices[0].DevEUI, Delete)},
+					Validators: []ValidatorFunc{ValidateDeviceQueueAccess(devices[0].DevEUI, Create), ValidateDeviceQueueAccess(devices[0].DevEUI, Read), ValidateDeviceQueueAccess(devices[0].DevEUI, List), ValidateDeviceQueueAccess(devices[0].DevEUI, Update), ValidateDeviceQueueAccess(devices[0].DevEUI, Delete)},
 					Claims:     Claims{Username: "user4"},
 					ExpectedOK: false,
 				},