+++ docs/content/overview/changelog.md
@@ -16,7 +16,8 @@ menu:
   *Add gateway meta-data* is disabled in the service-profile
   (thanks [@pni-jmattison](https://forum.loraserver.io/u/pni-jmattison/summary)).
 * Rename `/api/node/...` prefix of downlink queue into `/api/device/...`
-  (thanks [@iegomez](https://github.com/iegomez))
+  (thanks [@iegomez](https://github.com/iegomez)).
+* Rename `DownlinkQueue...` gRPC methods and structs into `DeviceQueue...`.
 
 ### 0.14.0